+++ test/shellcheck.hs
@@ -6,8 +6,10 @@ import qualified ShellCheck.Analytics
 import qualified ShellCheck.AnalyzerLib
 import qualified ShellCheck.Checker
 import qualified ShellCheck.Checks.Commands
+import qualified ShellCheck.Checks.Custom
 import qualified ShellCheck.Checks.ShellSupport
 import qualified ShellCheck.Fixer
+import qualified ShellCheck.Formatter.Diff
 import qualified ShellCheck.Parser
 
 main = do
@@ -17,8 +19,10 @@ main = do
         ,ShellCheck.AnalyzerLib.runTests
         ,ShellCheck.Checker.runTests
         ,ShellCheck.Checks.Commands.runTests
+        ,ShellCheck.Checks.Custom.runTests
         ,ShellCheck.Checks.ShellSupport.runTests
         ,ShellCheck.Fixer.runTests
+        ,ShellCheck.Formatter.Diff.runTests
         ,ShellCheck.Parser.runTests
       ]
     if and results